Add files via upload
The player's inventory in your game will be able to hold multiple items. You will need to use a data type that can do this. An array can be used to hold multiple values. For example:

let order = ["first", "second", "third"];
This is an array which holds three values. Notice how the values are separated by commas.

Change your inventory variable to be an array with the strings "stick", "dagger", and "sword".